complaint_system/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
drf-yasg>=1.21.7
gunicorn>=21.2.0
whitenoise>=6.6.0
